CTAKES-330 - autoload pipeline if it exists to save new user a step
git-svn-id: https://svn.apache.org/repos/asf/ctakes/trunk@1789252 13f79535-47bb-0310-9956-ffa450edef68

+++ ctakes-distribution/src/main/bin/runctakesCVD.bat
@@ -41,7 +41,14 @@ goto end
 @REM use JAVA_HOME if set
 if exist "%JAVA_HOME%\bin\java.exe" set PATH=%JAVA_HOME%\bin;%PATH%
 @set PATH=%PATH%;%CTAKES_HOME%\lib\auth\x64
+
 cd %CTAKES_HOME%
-java -cp "%CTAKES_HOME%/desc/;%CTAKES_HOME%/resources/;%CTAKES_HOME%/lib/*" -Dlog4j.configuration=file:/%CTAKES_HOME%/config/log4j.xml -Xms512M -Xmx3g org.apache.uima.tools.cvd.CVD
+
+if exist "desc/ctakes-clinical-pipeline/desc/analysis_engine/AggregatePlaintextFastUMLSProcessor.xml" ( 
+  java -cp "%CTAKES_HOME%/desc/;%CTAKES_HOME%/resources/;%CTAKES_HOME%/lib/*" -Dlog4j.configuration=file:/%CTAKES_HOME%/config/log4j.xml -Xms512M -Xmx3g org.apache.uima.tools.cvd.CVD -desc desc/ctakes-clinical-pipeline/desc/analysis_engine/AggregatePlaintextFastUMLSProcessor.xml
+) else (
+  java -cp "%CTAKES_HOME%/desc/;%CTAKES_HOME%/resources/;%CTAKES_HOME%/lib/*" -Dlog4j.configuration=file:/%CTAKES_HOME%/config/log4j.xml -Xms512M -Xmx3g org.apache.uima.tools.cvd.CVD
+)
+  
 cd %CURRENT_DIR%
 :end
